task3/services/service-a/requirements.txt
Flask
requests
opentelemetry-api  #api otel
opentelemetry-sdk #sdk otel
opentelemetry-instrumentation-flask #framework instrumentation
opentelemetry-instrumentation-requests #framework instrumentation
opentelemetry-exporter-otlp-proto-http
